requirements.txt
Pillow
opencv-python
pycaw
pynput
pystray
pyinstaller